requirements.txt
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
protobuf==3.12.2
pycparser==2.20
PyJWT==1.4.2
pymongo==3.10.1
python-decouple==3.3
six==1.15.0
Werkzeug==1.0.1
